template

+++ src/strfunc.cc
@@ -2,7 +2,7 @@
 
 std::string ff::quiti(const std::string &text) {
     std::string buff="";
-    for(int c=0;c<text.length();c++){
+    for(int c=0;c<(int)text.length();c++){
         if(text[c] != 'i' && text[c] != 'j' && text[c] != '*') {
             buff += text[c];
         }
@@ -12,7 +12,7 @@ std::string ff::quiti(const std::string &text) {
 
 std::string ff::trim(const std::string &text) {
     std::string buff="";
-    for(int c=0;c<text.length();c++){
+    for(int c=0;c<(int)text.length();c++){
         if(text[c] != ' ') {
             buff += text[c];
         }
